+++ R/dependencies.R
@@ -1066,6 +1066,7 @@ renv_dependencies_discover_r <- function(path  = NULL,
     renv_dependencies_discover_r_library_require,
     renv_dependencies_discover_r_require_namespace,
     renv_dependencies_discover_r_colon,
+    renv_dependencies_discover_r_citation,
     renv_dependencies_discover_r_pacman,
     renv_dependencies_discover_r_modules,
     renv_dependencies_discover_r_import,
@@ -1225,12 +1226,31 @@ renv_dependencies_discover_r_colon <- function(node, envir) {
   if (is.symbol(package))
     package <- as.character(package)
 
-  if (!is.character(package) || length(package) != 1)
+  if (!is.character(package) || length(package) != 1L)
     return(FALSE)
 
   envir[[package]] <- TRUE
   TRUE
 
+}
+
+renv_dependencies_discover_r_citation <- function(node, envir) {
+  
+  node <- renv_call_expect(node, "utils", "citation")
+  if (is.null(node))
+    return(FALSE)
+  
+  matched <- catch(match.call(utils::citation, node))
+  if (inherits(matched, "error"))
+    return(FALSE)
+  
+  package <- matched[["package"]]
+  if (!is.character(package) || length(package) != 1L)
+    return(FALSE)
+  
+  envir[[package]] <- TRUE
+  TRUE
+  
 }
 
 renv_dependencies_discover_r_pacman <- function(node, envir) {